% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{frecuencia_mortalidad}
\alias{frecuencia_mortalidad}
\title{Distribucion Exponencial}
\usage{
frecuencia_mortalidad(x, tiempo)
}
\arguments{
\item{x}{(vector) datos de la variable explicativa}

\item{y}{(vector) dato de la respuesta}
}
\value{
Una grafica
}
\description{
Realiza una Distribucion Exponencial
}
